YARN-11675. Update MemoryResourceHandler implementation for cgroup v2 support
